Implemet Stations Windows. TODO: Add handle of removings stations and drones.

+++ BL/DroneBL.cs
@@ -25,7 +25,7 @@ namespace BO
             _drone = idalObj.GetDroneById(droneId);
         }
 
-        public double Id => _drone.Id;
+        public int Id => _drone.Id;
 
         public string Model => _drone.Model;
